Move MySQL tests into separate directory
