+++ tests/Integration/Database/DatabaseTestCase.php
@@ -6,6 +6,13 @@ use Orchestra\Testbench\TestCase;
 
 abstract class DatabaseTestCase extends TestCase
 {
+    /**
+     * The current database driver.
+     *
+     * @return string
+     */
+    protected $driver;
+
     protected function getEnvironmentSetUp($app)
     {
         $app['config']->set('database.connections.testbench', [
@@ -17,6 +24,10 @@ abstract class DatabaseTestCase extends TestCase
         if (! env('DB_CONNECTION')) {
             $app['config']->set('database.default', 'testbench');
         }
+
+        $connection = $app['config']->get('database.default');
+
+        $this->driver = $app['config']->get("database.connections.$connection.driver");
     }
 
     protected function tearDown(): void
