+++ tests/Integration/Database/MySql/MySqlTestCase.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Database\MySql;
+
+use Illuminate\Tests\Integration\Database\DatabaseTestCase;
+
+abstract class MySqlTestCase extends DatabaseTestCase
+{
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        if ($this->driver !== 'mysql') {
+            $this->markTestSkipped('Test requires a MySQL connection.');
+        }
+    }
+}
